fix: axios e query na feature de agendamento
