+++ frontend/atendimento-app/src/features/agenda/hooks/useCriarAgendamento.ts
@@ -1,0 +1,15 @@
+import { useMutation, useQueryClient } from "@tanstack/react-query";
+import { criarAgendamento } from "../services/agendaService";
+
+export function useCriarAgendamento() {
+  const queryClient = useQueryClient();
+
+  return useMutation({
+    mutationFn: criarAgendamento,
+    onSuccess: (_, variables) => {
+      queryClient.invalidateQueries({
+        queryKey: ["agendamentos", variables.profissionalId],
+      });
+    },
+  });
+}
