+++ frontend/atendimento-app/src/proxy.ts
@@ -1,7 +1,7 @@
 import { NextResponse } from "next/server";
 import type { NextRequest } from "next/server";
 
-export function middleware(request: NextRequest) {
+export function proxy(request: NextRequest) {
   const verified = request.cookies.get("verified")?.value;
 
   const publicRoutes = ["/login"];
